+++ src/main/gen/com/windea/plugin/idea/paradox/localisation/psi/impl/ParadoxLocalisationSequentialNumberImpl.java
@@ -1,23 +1,23 @@
 // This is a generated file. Not intended for manual editing.
 package com.windea.plugin.idea.paradox.localisation.psi.impl;
 
-import java.util.List;
 import org.jetbrains.annotations.*;
 import com.intellij.lang.ASTNode;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiElementVisitor;
-import com.intellij.psi.util.PsiTreeUtil;
+
 import static com.windea.plugin.idea.paradox.localisation.psi.ParadoxLocalisationTypes.*;
 import com.windea.plugin.idea.paradox.localisation.psi.*;
 
-public class ParadoxLocalisationSerialNumberImpl extends ParadoxLocalisationNamedElementImpl implements ParadoxLocalisationSerialNumber {
+public class ParadoxLocalisationSequentialNumberImpl extends ParadoxLocalisationNamedElementImpl implements
+    ParadoxLocalisationSequentialNumber {
 
-  public ParadoxLocalisationSerialNumberImpl(@NotNull ASTNode node) {
+  public ParadoxLocalisationSequentialNumberImpl(@NotNull ASTNode node) {
     super(node);
   }
 
   public void accept(@NotNull ParadoxLocalisationVisitor visitor) {
-    visitor.visitSerialNumber(this);
+    visitor.visitSequentialNumber(this);
   }
 
   @Override
@@ -28,8 +28,8 @@ public class ParadoxLocalisationSerialNumberImpl extends ParadoxLocalisationName
 
   @Override
   @Nullable
-  public PsiElement getSerialNumberId() {
-    return findChildByType(SERIAL_NUMBER_ID);
+  public PsiElement getSequentialNumberId() {
+    return findChildByType(SEQUENTIAL_NUMBER_ID);
   }
 
   @Override