+++ src/main/kotlin/com/windea/plugin/idea/paradox/localisation/highlighter/ParadoxLocalisationSyntaxHighlighter.kt
@@ -21,7 +21,7 @@ class ParadoxLocalisationSyntaxHighlighter : SyntaxHighlighterBase() {
 		private val PARAMETER_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.PARAMETER_KEY)
 		private val PROPERTY_REFERENCE_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.PROPERTY_REFERENCE_KEY)
 		private val ICON_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.ICON_KEY)
-		private val SERIAL_NUMBER_ID_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.SERIAL_NUMBER_KEY)
+		private val SEQUENTIAL_NUMBER_ID_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.SEQUENTIAL_NUMBER_KEY)
 		private val COLOR_ID_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.COLOR_KEY)
 		private val COMMAND_SCOPE_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.COMMAND_SCOPE_KEY)
 		private val COMMAND_FIELD_KEYS = arrayOf(ParadoxLocalisationAttributesKeys.COMMAND_FIELD_KEY)
@@ -39,7 +39,7 @@ class ParadoxLocalisationSyntaxHighlighter : SyntaxHighlighterBase() {
 		STRING_TOKEN, LEFT_QUOTE, RIGHT_QUOTE -> STRING_KEYS
 		PROPERTY_REFERENCE_START, PARAMETER_SEPARATOR, PROPERTY_REFERENCE_END,
 		ICON_START, ICON_END,
-		SERIAL_NUMBER_START, SERIAL_NUMBER_END, 
+		SEQUENTIAL_NUMBER_START, SEQUENTIAL_NUMBER_END, 
 		COLORFUL_TEXT_START, COLORFUL_TEXT_END,
 		COMMAND_START, COMMAND_END -> MARKER_KEYS
 		PROPERTY_REFERENCE_PARAMETER, ICON_PARAMETER -> PARAMETER_KEYS
@@ -47,7 +47,7 @@ class ParadoxLocalisationSyntaxHighlighter : SyntaxHighlighterBase() {
 		COMMAND_SCOPE -> COMMAND_SCOPE_KEYS
 		COMMAND_FIELD -> COMMAND_FIELD_KEYS
 		ICON_ID -> ICON_KEYS
-		SERIAL_NUMBER_ID -> SERIAL_NUMBER_ID_KEYS
+		SEQUENTIAL_NUMBER_ID -> SEQUENTIAL_NUMBER_ID_KEYS
 		COLOR_ID -> COLOR_ID_KEYS
 		COMMENT, END_OF_LINE_COMMENT, ROOT_COMMENT -> COMMENT_KEYS
 		VALID_ESCAPE_TOKEN -> VALID_ESCAPE_KEYS