+++ src/main/kotlin/com/windea/plugin/idea/paradox/localisation/intentions/ChangeSerialNumberIntention.kt
@@ -10,9 +10,9 @@ import com.intellij.psi.*
 import com.windea.plugin.idea.paradox.*
 import com.windea.plugin.idea.paradox.localisation.psi.*
 
-object ChangeSerialNumberIntention : IntentionAction {
-	private val _name = message("paradox.localisation.intention.changeSerialNumber")
-	private val _title = message("paradox.localisation.intention.changeSerialNumber.title")
+object ChangeSequentialNumberIntention : IntentionAction {
+	private val _name = message("paradox.localisation.intention.changeSequentialNumber")
+	private val _title = message("paradox.localisation.intention.changeSequentialNumber.title")
 	
 	override fun startInWriteAction() = false
 	
@@ -23,28 +23,28 @@ object ChangeSerialNumberIntention : IntentionAction {
 	override fun isAvailable(project: Project, editor: Editor?, file: PsiFile?): Boolean {
 		if(editor == null || file == null) return false
 		val element = file.findElementAt(editor.caretModel.offset)?.parent
-		return element is ParadoxLocalisationSerialNumber
+		return element is ParadoxLocalisationSequentialNumber
 	}
 	
 	override fun invoke(project: Project, editor: Editor?, file: PsiFile?) {
 		if(editor == null || file == null) return
 		val element = file.findElementAt(editor.caretModel.offset)?.parent
-		if(element is ParadoxLocalisationSerialNumber) {
-			JBPopupFactory.getInstance().createListPopup(Popup(element, paradoxSerialNumbers)).showInBestPositionFor(editor)
+		if(element is ParadoxLocalisationSequentialNumber) {
+			JBPopupFactory.getInstance().createListPopup(Popup(element, paradoxSequentialNumbers)).showInBestPositionFor(editor)
 		}
 	}
 	
 	private class Popup(
-		private val value: ParadoxLocalisationSerialNumber,
-		values: Array<ParadoxSerialNumber>
-	) : BaseListPopupStep<ParadoxSerialNumber>(_title, *values) {
-		override fun getTextFor(value: ParadoxSerialNumber) = value.popupText
+		private val value: ParadoxLocalisationSequentialNumber,
+		values: Array<ParadoxSequentialNumber>
+	) : BaseListPopupStep<ParadoxSequentialNumber>(_title, *values) {
+		override fun getTextFor(value: ParadoxSequentialNumber) = value.popupText
 		
 		override fun getDefaultOptionIndex() = 0
 		
 		override fun isSpeedSearchEnabled(): Boolean = true
 		
-		override fun onChosen(selectedValue: ParadoxSerialNumber, finalChoice: Boolean): PopupStep<*>? {
+		override fun onChosen(selectedValue: ParadoxSequentialNumber, finalChoice: Boolean): PopupStep<*>? {
 			//需要在WriteCommandAction里面执行
 			runWriteCommandAction(value.project) { value.name = selectedValue.name }
 			return PopupStep.FINAL_CHOICE